// src/domain/contracts/repos/schema.ts
type TNFeInfNFeDetProdIndEscala = 'S' | 'N';
// ...
export interface TNFeInfNFeDet {
  $: {
    nItem: string;
  };
  prod: TNFeInfNFeDetProd;
  imposto: TNFeInfNFeDetImposto;
  impostoDevol?: TNFeInfNFeDetImpostoDevol;
  infAdProd?: string;
}
interface TNFeInfNFeDetImpostoDevol {
  pDevol: number;
  IPI: TNFeInfNFeDetImpostoDevolIPI;
}
interface TNFeInfNFeDetImpostoDevolIPI {
  vIPIDevol: number;
}
interface TNFeInfNFeDetImpostoPIS {
  PISAliq?: TNFeInfNFeDetImpostoPISItem;
  PISQtde?: TNFeInfNFeDetImpostoPISItem;
  PISNT?: TNFeInfNFeDetImpostoPISItem;
  PISOutr?: TNFeInfNFeDetImpostoPISItem;
}
interface TNFeInfNFeDetImpostoPISItem {
  CST: string;
  vBC?: number;
  pPIS?: number;
  vPIS?: number;
  vBCProd?: number;
  qBCProd?: number;
  vAliqProd?: number;
}
interface TNFeInfNFeDetImposto {
  vTotTrib?: number;
  ICMS: TNFeInfNFeDetImpostoICMS;
  PIS: TNFeInfNFeDetImpostoPIS;
  COFINS: TNFeInfNFeDetImpostoCOFINS;
  IPI?: TNFeInfNFeDetImpostoIPI;
  II?: TNFeInfNFeDetImpostoII;
  ISSQN?: TNFeInfNFeDetImpostoISSQN;
  PISST?: TNFeInfNFeDetImpostoPISST;
  COFINSST?: TNFeInfNFeDetImpostoCOFINSST;
  ICMSUFDest?: TNFeInfNFeDetImpostoICMSUFDest;
}
interface TNFeInfNFeDetImpostoICMSUFDest {
  vBCUFDest: number;
  vBCFCPUFDest: number;
  pFCPUFDest: number;
  pICMSUFDest: number;
  pICMSInter: number;
  pICMSInterPart: number;
  vFCPUFDest: number;
  vICMSUFDest: number;
  vICMSUFRemet: number;
}
interface TNFeInfNFeDetImpostoCOFINSST {
  vBC?: number;
  pCOFINS?: number;
  vCOFINS: number;
  qBCProd?: number;
  vAliqProd?: number;
}
interface TNFeInfNFeDetImpostoPISST {
  vBC?: number;
  pPIS?: number;
  vPIS?: number;
  qBCProd?: number;
  vAliqProd?: number;
}
interface TNFeInfNFeDetImpostoISSQN {
  vBC: string;
  vAliq: string;
  vISSQN: string;
  cMunFG: string;
  cListServ: TCListServ;
  vDeducao: string;
  vOutro: string;
  vDescIncond: string;
  vDescCond: string;
  vISSRet: string;
  indISS: TNFeInfNFeDetImpostoISSQNIndISS;
  cServico: string;
  cMun: string;
  cPais: string;
  nProcesso: string;
  indIncentivo: TNFeInfNFeDetImpostoISSQNIndIncentivo;
}
interface TNFeInfNFeDetImpostoII {
  vBC: number;
  vDespAdu: number;
  vII: number;
  vIOF: number;
}
interface TNFeInfNFeDetImpostoIPI {
  cEnq: string;
  CNPJProd: string;
  cSelo: string;
  qSelo: string;
  IPINT?: string;
  IPITRIB?: TIpiIPITrib;
}
interface TIpiIPITrib {
  CST: TIpiIPITribCST;
  vBC?: number;
  pIPI?: number;
  vIPI: number;
  qUnid?: number;
  vUnid?: number;
}
interface TNFeInfNFeDetImpostoCOFINS {
  COFINSAliq?: TNFeInfNFeDetImpostoCOFINSItem;
  COFINSQtde?: TNFeInfNFeDetImpostoCOFINSItem;
  COFINSNT?: TNFeInfNFeDetImpostoCOFINSItem;
  COFINSOutr?: TNFeInfNFeDetImpostoCOFINSItem;
}
interface TNFeInfNFeDetImpostoCOFINSItem {
  CST: string;
  vBC?: number;
  pCOFINS?: number;
  vCOFINS?: number;
  qBCProd?: number;
  vAliqProd?: number;
}
interface TNFeInfNFeDetImpostoICMS {
  ICMS00?: TNFeInfNFeDetImpostoICMSICMS00;
  ICMS10?: TNFeInfNFeDetImpostoICMSICMS10 | TNFeInfNFeDetImpostoICMSICMSPart;
  ICMS20?: TNFeInfNFeDetImpostoICMSICMS20;
  ICMS30?: TNFeInfNFeDetImpostoICMSICMS30;
  ICMS40?: TNFeInfNFeDetImpostoICMSICMS40;
  ICMS41?: TNFeInfNFeDetImpostoICMSICMSST;
  ICMS51?: TNFeInfNFeDetImpostoICMSICMS51;
  ICMS60?: TNFeInfNFeDetImpostoICMSICMS60;
  ICMS61?: TNFeInfNFeDetImpostoICMSICMS61;
  ICMS70?: TNFeInfNFeDetImpostoICMSICMS70;
  ICMS90?: TNFeInfNFeDetImpostoICMSICMS90 | TNFeInfNFeDetImpostoICMSICMSPart;
  ICMSPart?: TNFeInfNFeDetImpostoICMSICMSPart;
  ICMSSN101?: TNFeInfNFeDetImpostoICMSICMSSN101;
  ICMSSN102?: TNFeInfNFeDetImpostoICMSICMSSN102;
  ICMSSN201?: TNFeInfNFeDetImpostoICMSICMSSN201;
  ICMSSN202?: TNFeInfNFeDetImpostoICMSICMSSN202;
  ICMSSN500?: TNFeInfNFeDetImpostoICMSICMSSN500;
  ICMSSN900?: TNFeInfNFeDetImpostoICMSICMSSN900;
  ICMSST?: TNFeInfNFeDetImpostoICMSICMSST;
}
// ...
interface TNFeInfNFeDetProd {
  cProd: string;
  cEAN: string;
  xProd: string;
  NCM: string;
  nVE?: string[];
  CEST?: string;
  indEscala?: TNFeInfNFeDetProdIndEscala;
  indEscalaSpecified?: boolean;
  cNPJFab?: string;
  cBenef?: string;
  eXTIPI?: string;
  CFOP: string;
  uCom: string;
  qCom: string;
  vUnCom: string;
  vProd: string;
  cEANTrib: string;
  uTrib: string;
  qTrib: string;
  vUnTrib: string;
  vFrete: string;
  vSeg: string;
  vDesc: string;
  vOutro: string;
  indTot: '0' | '1';
  di?: TNFeInfNFeDetProdDI[];
  det?: TNFeInfNFeDetProdDet[];
  xPed?: string;
  nItemPed?: string;
  nFCI?: string;
  rastro?: TNFeInfNFeDetProdRastro[];
  comb?: TNFeInfNFeDetProdComb;
  items?: object[];
}
interface TNFeInfNFeDetProdRastro {
  nLote: string;
  qLote: string;
  dFab: string;
  dVal: string;
  cAgreg: string;
}
export interface TNFeInfNFeDetProdComb {
  cProdANP: string;
  descANP: string;
  pGLP_Opc?: number;
  pGNn_Opc?: number;
  pGNi_Opc?: number;
  vPart_Opc?: number;
  CODIF_Opc?: string;
  qTemp_Opc?: number;
  UFCons: string;
  qBCprod_Opc?: number;
  vAliqProd_Opc?: number;
  vCIDE_Opc?: number;
  encerrante_Opc?: string;
  encerrante: TNFeInfNFeDetProdCombEncerrante;
}
export interface TNFeInfNFeDetProdCombEncerrante {
  nBico: string;
  nBomba: string;
  nTanque: string;
  vEncIni: string;
  vEncFin: string;
}
interface TNFeInfNFeDetProdDet {
  nDraw: string;
  Ind: TNFeInfNFeDetProdDetInd;
}
interface TNFeInfNFeDetProdDetInd {
  nRE: string;
  chNFe: string;
  q: string;
}
interface TNFeInfNFeDetProdDI {
  nDI: string;
  dDI: string;
  xLocDesemb: string;
  uFDesemb: TUfEmi;
  dDesemb: string;
  tpViaTransp: TNFeInfNFeDetProdDITpViaTransp;
  vAFRMM: string;
  tpIntermedio: TNFeInfNFeDetProdDITpIntermedio;
  cNPJ: string;
  uFTerceiro: TUfEmi;
  uFTerceiroSpecified: boolean;
  cador: string;
  adi: TNFeInfNFeDetProdDIAdi[];
}